requirements.txt
fastapi==0.128.1
idc_index==0.11.7
numpy==2.4.2
Pillow==12.1.0
protobuf==6.33.5
pydicom==3.0.1
google-cloud-storage==3.9.0